WI{ refacto tile size COLLISION BUG

+++ cub3d.h
@@ -6,7 +6,7 @@
 /*   By: lpassera <lpassera@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/15 15:06:26 by lpassera          #+#    #+#             */
-/*   Updated: 2021/01/18 16:31:09 by lpassera         ###   ########.fr       */
+/*   Updated: 2021/01/19 10:34:33 by lpassera         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void set_ray_direction(t_ray *ray);
 void cast_rays(t_game *game);
 void render_rays(t_game *game);
 void render_3d_walls(t_game *game);
-t_bool has_wall_at(t_game *game, double x, double y);
+t_bool has_wall_at(t_game *game, float x, float y);
 int main_loop(t_game *game);
 void set_pos(t_pos *pos, float x, float y);
 float pos_distance(t_pos *p1, t_pos *p2);
